fix #2: update client with security token

+++ client/model_defect_create.go
@@ -4,7 +4,7 @@ type DefectCreate struct {
 	Title        string   `json:"title"`
 	ActualResult string   `json:"actual_result"`
 	Severity     int32    `json:"severity"`
-	MilestoneId  int32    `json:"milestone_id,omitempty"`
+	MilestoneId  int64    `json:"milestone_id,omitempty"`
 	Attachments  []string `json:"attachments,omitempty"`
 	// A map of custom fields values (id => value)
 	CustomField map[string]string `json:"custom_field,omitempty"`
