+++ client/model_inline_response_200_result.go
@@ -1,6 +1,6 @@
 package client
 
 type InlineResponse200Result struct {
-	CaseId int32  `json:"case_id,omitempty"`
+	CaseId int64  `json:"case_id,omitempty"`
 	Hash   string `json:"hash,omitempty"`
 }
