+++ client/model_plan_create.go
@@ -3,5 +3,5 @@ package client
 type PlanCreate struct {
 	Title       string  `json:"title"`
 	Description string  `json:"description,omitempty"`
-	Cases       []int32 `json:"cases"`
+	Cases       []int64 `json:"cases"`
 }
