+++ client/model_plan_detailed.go
@@ -5,7 +5,7 @@ import (
 )
 
 type PlanDetailed struct {
-	Id          int32               `json:"id,omitempty"`
+	Id          int64               `json:"id,omitempty"`
 	Title       string              `json:"title,omitempty"`
 	Description string              `json:"description,omitempty"`
 	CasesCount  int32               `json:"cases_count,omitempty"`
