+++ client/model_plan_update.go
@@ -3,5 +3,5 @@ package client
 type PlanUpdate struct {
 	Title       string  `json:"title,omitempty"`
 	Description string  `json:"description,omitempty"`
-	Cases       []int32 `json:"cases,omitempty"`
+	Cases       []int64 `json:"cases,omitempty"`
 }
