+++ client/model_project_access.go
@@ -2,5 +2,5 @@ package client
 
 type ProjectAccess struct {
 	// Team member id title.
-	MemeberId int32 `json:"memeber_id,omitempty"`
+	MemeberId int64 `json:"memeber_id,omitempty"`
 }
