+++ client/model_result_update.go
@@ -2,7 +2,7 @@ package client
 
 type ResultUpdate struct {
 	Status      string              `json:"status,omitempty"`
-	TimeMs      int32               `json:"time_ms,omitempty"`
+	TimeMs      int64               `json:"time_ms,omitempty"`
 	Defect      bool                `json:"defect,omitempty"`
 	Attachments []string            `json:"attachments,omitempty"`
 	Stacktrace  string              `json:"stacktrace,omitempty"`
